부산대 BE 송승우 2단계 - 회원 로그인 (#229)
* docs: 1단계 유효성 검사 및 예외 처리 개발 예정 기능 목록 추가

* refactor: 상품 요청 DTO 공통 필드를 ProductRequestBase로 분리

* refactor: 커스텀 어노테이션을 적용하여 상품 이름, 상품 가격 검증

* refactor: 상품 이름에 '카카오' 포함 시 예외 처리 로직 엔티티 내부로 이동

* feat: Product에 유효성 검증 적용 및 테스트 코드 작성

* docs: 2단계 개발 예정 기능 목록 추가(회원 기능)

* feat: 회원 테이블 생성

* feat: 회원가입 API 구현

* feat: 로그인 API 구현

* feat: member 테스트 코드 작성

* docs: 개발 예정 기능 목록 - 인증 및 예외처리 추가

* feat: Authorization 헤더 또는 토큰이 유효하지 않은 경우 예외처리

* feat: 인가에 대한 검증 기능 추가

* refactor: 로그인 로직 수정

+++ src/main/java/gift/controller/ProductController.java
@@ -1,5 +1,6 @@
 package gift.controller;
 
+import gift.component.JwtUtil;
 import gift.dto.CreateProductRequestDto;
 import gift.dto.ProductResponseDto;
 import gift.dto.UpdateProductRequestDto;
@@ -15,9 +16,11 @@ import java.util.List;
 @RequestMapping("api/products")
 public class ProductController {
     private final ProductService productService;
+    private final JwtUtil jwtUtil;
 
-    public ProductController(ProductService productService) {
+    public ProductController(ProductService productService, JwtUtil jwtUtil) {
         this.productService = productService;
+        this.jwtUtil = jwtUtil;
     }
 
     // 상품 목록 조회
@@ -38,7 +41,10 @@ public class ProductController {
 
     // 상품 추가
     @PostMapping
-    public ResponseEntity<ProductResponseDto> createProduct(@Valid @RequestBody CreateProductRequestDto requestDto) {
+    public ResponseEntity<ProductResponseDto> createProduct(
+            @RequestHeader(value = "Authorization", required = false) String authHeader,
+            @Valid @RequestBody CreateProductRequestDto requestDto) {
+        jwtUtil.validateAuthorizationHeader(authHeader, "products-api");
         ProductResponseDto productResponseDto = productService.createProduct(requestDto);
 
         return new ResponseEntity<>(productResponseDto, HttpStatus.CREATED);
@@ -46,7 +52,10 @@ public class ProductController {
 
     // 상품 삭제
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteProduct(
+            @RequestHeader(value = "Authorization", required = false) String authHeader,
+            @PathVariable Long id) {
+        jwtUtil.validateAuthorizationHeader(authHeader, "products-api");
         productService.deleteProduct(id);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -55,9 +64,11 @@ public class ProductController {
     // 상품 전체 수정
     @PutMapping("/{id}")
     public ResponseEntity<Void> updateProduct(
+            @RequestHeader(value = "Authorization", required = false) String authHeader,
             @PathVariable Long id,
             @Valid @RequestBody UpdateProductRequestDto requestDto
     ) {
+        jwtUtil.validateAuthorizationHeader(authHeader, "products-api");
         productService.updateProduct(requestDto);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
